separate round_end
