+++ include/game/round_end.h
@@ -1,0 +1,8 @@
+#ifndef GAME_ROUND_END_H
+#define GAME_ROUND_END_H
+
+void game_round_end_on_init(void);
+void game_round_end_on_update(void);
+void game_round_end_on_exit(void);
+int calculate_interest_reward(void);
+#endif // GAME_ROUND_END_H
