+++ source/game.c
@@ -15,6 +15,7 @@
 #include "game/point.h"
 #include "game/rect.h"
 #include "game/round.h"
+#include "game/round_end.h"
 #include "game/selection.h"
 #include "game/timer.h"
 #include "graphic_utils.h"
@@ -118,8 +119,8 @@ static void noop(void)
 // reviewer(s).
 extern void game_round_on_init(void);
 extern void game_playing_on_update(void);
-static void game_round_end_on_update(void);
-static void game_round_end_on_exit(void);
+extern void game_round_end_on_update(void);
+extern void game_round_end_on_exit(void);
 static void game_shop_on_update(void);
 static void game_shop_on_exit(void);
 static void game_lose_on_init(void);
@@ -130,29 +131,19 @@ static void game_win_on_update(void);
 static void game_shop_intro(void);
 static void game_shop_process_user_input(void);
 static void game_shop_outro(void);
-static void game_round_end_start(void);
-static void game_round_end_start_expand_popup(void);
-static void game_round_end_display_finished_blind(void);
-static void game_round_end_display_score_min(void);
-static void game_round_end_update_blind_reward(void);
-static void game_round_end_panel_exit(void);
-static void game_round_end_display_rewards(void);
-static void game_round_end_display_cashout(void);
-static void game_round_end_dismiss_round_end_panel(void);
 
 void change_background(enum BackgroundId id);
 void display_temp_score(u32 value);
 void display_score(u32 value);
 static void check_flaming_score(void);
 void display_round(int value);
-static void display_hands(int value);
-static void display_discards(int value);
+void display_hands(int value);
+void display_discards(int value);
 void set_hand(void);
 int deck_get_size(void);
 int deck_get_max_size(void);
 void increment_blind(enum BlindState increment_reason);
 static void game_over_init(void);
-static int calculate_interest_reward(void);
 static void game_over_anim_frame(void);
 
 static void game_playing_discard_on_pressed(void);
@@ -208,7 +199,6 @@ static void game_shop_create_items(void);
 
 static void erase_price_under_sprite_object(SpriteObject* sprite_object);
 static void print_price_under_sprite_object(SpriteObject* sprite_object, int price);
-static void game_round_end_extend_black_panel_down(int black_panel_bottom);
 
 static void remove_owned_joker(int owned_joker_idx);
 
@@ -317,18 +307,6 @@ static const SubStateActionFn shop_state_actions[] = {
     game_shop_outro
 };
 
-static const SubStateActionFn round_end_state_actions[] = {
-    game_round_end_start,
-    game_round_end_start_expand_popup,
-    game_round_end_display_finished_blind,
-    game_round_end_display_score_min,
-    game_round_end_update_blind_reward,
-    game_round_end_panel_exit,
-    game_round_end_display_rewards,
-    game_round_end_display_cashout,
-    game_round_end_dismiss_round_end_panel
-};
-
 static int reroll_cost = REROLL_BASE_COST;
 
 // The current game state, this is used to determine what the game is doing at any given time
@@ -357,22 +335,16 @@ enum BlindState blinds_states[BLIND_TYPE_MAX] = {
     BLIND_STATE_UPCOMING
 };
 
-static int blind_reward = 0;
-static int hand_reward = 0;
-static int interest_reward = 0;
-static int interest_to_count = 0;
-static int interest_start_time = UNDEFINED;
-
 // Red deck default (can later be moved to a deck.h file or something)
-static int max_hands = 4;
-static int max_discards = 4;
+int max_hands = 4;
+int max_discards = 4;
 // Set in game_init and game_round_init
 int hands = 0;
-static int discards = 0;
+int discards = 0;
 
 int game_round = 0;
 int ante = 0;
-static int money = 0;
+int money = 0;
 u32 score = 0;
 u32 temp_score = 0; // This is the score that shows in the same spot as the hand type.
 bool score_flames_active = false;
@@ -997,14 +969,6 @@ bool shift_null_card_to_end(int null_card_idx)
     return true;
 }
 
-/* Copies the appropriate item into the top left panel (blind/shop icon)
- * from where it was put outside the screenview
- */
-static void bg_copy_current_item_to_top_left_panel(void)
-{
-    main_bg_se_copy_rect(TOP_LEFT_ITEM_SRC_RECT, TOP_LEFT_PANEL_POINT);
-}
-
 // Resets bottom row bg tiles of the top left panel (shop/blind) after
 // it is dismissed to match the rest of the game panel background.
 void reset_top_left_panel_bottom_row()
@@ -1023,84 +987,11 @@ void change_background(enum BackgroundId id)
     }
     else if (id == BG_CARD_SELECTING)
     {
-        tte_erase_rect_wrapper(HAND_SIZE_RECT_PLAYING);
-        REG_WIN0V = (REG_WIN0V << 8) | 0x80; // Set window 0 top to 128
-
-        if (background == BG_CARD_PLAYING)
-        {
-            int offset = 11;
-            memcpy16(
-                &se_mem[MAIN_BG_SBB][SE_ROW_LEN * offset],
-                &background_gfxMap[SE_ROW_LEN * offset],
-                SE_ROW_LEN * 8
-            );
-        }
-        else
-        {
-            toggle_windows(true, true); // Enable window 0 for the hand shadow
-
-            // Load the tiles and palette
-            // Background
-            GRIT_CPY(pal_bg_mem, background_gfxPal);
-            GRIT_CPY(&tile8_mem[MAIN_BG_CBB], background_gfxTiles);
-            GRIT_CPY(&se_mem[MAIN_BG_SBB], background_gfxMap);
-
-            if (current_blind == BLIND_TYPE_BIG) // Change text and palette depending on blind type
-            {
-                main_bg_se_copy_rect(BIG_BLIND_TITLE_SRC_RECT, TOP_LEFT_BLIND_TITLE_POINT);
-            }
-            else if (current_blind == BLIND_TYPE_BOSS)
-            {
-                main_bg_se_copy_rect(BOSS_BLIND_TITLE_SRC_RECT, TOP_LEFT_BLIND_TITLE_POINT);
-
-                affine_background_set_color(
-                    blind_get_color(BLIND_TYPE_BOSS, BLIND_SHADOW_COLOR_INDEX)
-                );
-            }
-
-            bg_copy_current_item_to_top_left_panel();
-
-            // This would change the palette of the background to match the blind, but the backgroun
-            // doesn't use the blind token's exact colors so a different approach is required
-            memset16(
-                &pal_bg_mem[BLIND_BG_PRIMARY_PID],
-                blind_get_color(current_blind, BLIND_BACKGROUND_MAIN_COLOR_INDEX),
-                1
-            );
-            memset16(
-                &pal_bg_mem[BLIND_BG_SECONDARY_PID],
-                blind_get_color(current_blind, BLIND_BACKGROUND_SECONDARY_COLOR_INDEX),
-                1
-            );
-            memset16(
-                &pal_bg_mem[BLIND_BG_SHADOW_PID],
-                blind_get_color(current_blind, BLIND_BACKGROUND_SHADOW_COLOR_INDEX),
-                1
-            );
-
-            for (int i = 0; i < NUM_ELEM_IN_ARR(game_playing_buttons); i++)
-            {
-                button_set_highlight(&game_playing_buttons[i], false);
-            }
-        }
+        game_selecting_change_background(id);
     }
     else if (id == BG_CARD_PLAYING)
     {
-        if (background != BG_CARD_SELECTING)
-        {
-            change_background(BG_CARD_SELECTING);
-            background = BG_CARD_PLAYING;
-        }
-
-        REG_WIN0V = (REG_WIN0V << 8) | 0xA0; // Set window 0 bottom to 160
-        toggle_windows(true, true);
-
-        for (int i = 0; i <= 2; i++)
-        {
-            main_bg_se_move_rect_1_tile_vert(HAND_BG_RECT_SELECTING, SCREEN_DOWN);
-        }
-
-        tte_erase_rect_wrapper(HAND_SIZE_RECT_SELECT);
+        game_playing_change_background(id);
     }
     else if (id == BG_ROUND_END)
     {
@@ -1237,13 +1128,13 @@ void display_round(int value)
     );
 }
 
-static void display_hands(int value)
+void display_hands(int value)
 {
     // tte_erase_rect_wrapper(HANDS_TEXT_RECT);
     tte_printf("#{P:%d,%d; cx:0xD000}%d", HANDS_TEXT_RECT.left, HANDS_TEXT_RECT.top, hands); // Hand
 }
 
-static void display_discards(int value)
+void display_discards(int value)
 {
     // tte_erase_rect_wrapper(DISCARDS_TEXT_RECT);
     // Discard
@@ -1488,382 +1379,6 @@ static int game_playing_hand_row_get_size(void)
 int game_playing_button_row_get_size(void)
 {
     return NUM_ELEM_IN_ARR(game_playing_buttons);
-}
-
-static int calculate_interest_reward(void)
-{
-    int reward = (money / 5) * INTEREST_PER_5;
-    if (reward > MAX_INTEREST)
-        reward = MAX_INTEREST;
-    return reward;
-}
-
-static void game_round_end_on_exit()
-{
-    // Cleanup blind tokens from this round to avoid accumulating
-    // allocated blind sprites each round
-    blind_reward = 0;
-    hand_reward = 0;
-    interest_reward = 0;
-    sprite_destroy(&playing_blind_token);
-    sprite_destroy(&round_end_blind_token);
-    // TODO: Reuse sprites for blind selection?
-}
-
-static void game_round_end_on_update()
-{
-    if (state_info[game_state].substate == ROUND_END_EXIT)
-    {
-        game_change_state(GAME_STATE_SHOP);
-        return;
-    }
-
-    int substate = state_info[game_state].substate;
-    round_end_state_actions[substate]();
-}
-
-static void game_round_end_start()
-{
-    // Reset static variables to default values upon re-entering the round end state
-    if (timer == TM_RESET_STATIC_VARS)
-    {
-        change_background(BG_ROUND_END); // Change the background to the round end background
-        state_info[game_state].substate = START_EXPAND_POPUP; // Change the state to the next one
-        timer = TM_ZERO;                                      // Reset the timer
-        blind_reward = blind_get_reward(current_blind);
-        hand_reward = hands;
-        interest_reward = calculate_interest_reward();
-        interest_to_count = interest_reward;
-        interest_start_time = UNDEFINED;
-    }
-}
-
-static void game_round_end_start_expand_popup()
-{
-    main_bg_se_copy_rect_1_tile_vert(POP_MENU_ANIM_RECT, SCREEN_UP);
-
-    if (timer == TM_END_POP_MENU_ANIM)
-    {
-        state_info[game_state].substate = DISPLAY_FINISHED_BLIND;
-        timer = TM_ZERO;
-    }
-}
-
-static void game_round_end_extend_black_panel_down(int black_panel_bottom)
-{
-    Rect single_line_rect = ROUND_END_MENU_RECT;
-    single_line_rect.bottom = black_panel_bottom;
-    single_line_rect.top = single_line_rect.bottom - 1;
-    main_bg_se_copy_rect_1_tile_vert(single_line_rect, SCREEN_DOWN);
-}
-
-static void game_round_end_display_finished_blind()
-{
-    obj_unhide(round_end_blind_token->obj, 0);
-
-    int current_ante = ante;
-
-    // Beating the boss blind increases the ante, so we need to display the previous ante value
-    if (current_blind == BLIND_TYPE_BOSS)
-        current_ante--;
-
-    Rect blind_req_rect = ROUND_END_BLIND_REQ_RECT;
-    u32 blind_req = blind_get_requirement(current_blind, current_ante);
-
-    /* Not bothering to truncate here because there are 8 tiles
-     * and the blind requirement will not increase past ante 8
-     * so there's enough room for sure.
-     */
-    char blind_req_str_buff[UINT_MAX_DIGITS + 1];
-    snprintf(blind_req_str_buff, sizeof(blind_req_str_buff), "%lu", blind_req);
-
-    update_text_rect_to_right_align_str(&blind_req_rect, blind_req_str_buff, OVERFLOW_RIGHT);
-
-    tte_printf(
-        "#{P:%d,%d; cx:0x%X000}%s",
-        blind_req_rect.left,
-        blind_req_rect.top,
-        TTE_RED_PB,
-        blind_req_str_buff
-    );
-
-    if (timer == TM_START_ROUND_END_REWARDS_ANIM)
-    {
-        game_round_end_extend_black_panel_down(ROUND_END_BLACK_PANEL_INIT_BOTTOM_SE);
-    }
-
-    if (timer >= TM_END_DISPLAY_FIN_BLIND)
-    {
-        state_info[game_state].substate = DISPLAY_SCORE_MIN;
-        timer = TM_ZERO;
-    }
-}
-
-static void game_round_end_display_score_min()
-{
-    const int timer_offset = timer - 1;
-    const int x_from = 0;
-    const int y_from = 29;
-    const int x_to = 13;
-    const int y_to = 11;
-
-    memcpy16(
-        &se_mem[MAIN_BG_SBB][x_to + timer_offset + 32 * y_to],
-        &se_mem[MAIN_BG_SBB][x_from + timer_offset + 32 * y_from],
-        1
-    );
-
-    if (timer >= TM_END_DISPLAY_SCORE_MIN)
-    {
-        state_info[game_state].substate = UPDATE_BLIND_REWARD;
-        timer = TM_ZERO;
-    }
-}
-
-static void game_round_end_update_blind_reward()
-{
-    if (timer % FRAMES(20) != 0)
-        return;
-
-    // TODO: Add sound effect here
-
-    if (blind_reward > 0)
-    {
-        blind_reward--;
-        tte_printf(
-            "#{P:%d,%d; cx:0x%X000}$%d",
-            BLIND_REWARD_RECT.left,
-            BLIND_REWARD_RECT.top,
-            TTE_YELLOW_PB,
-            blind_reward
-        );
-        tte_printf(
-            "#{P:%d,%d; cx:0x%X000}$%d",
-            ROUND_END_BLIND_REWARD_RECT.left,
-            ROUND_END_BLIND_REWARD_RECT.top,
-            TTE_YELLOW_PB,
-            blind_get_reward(current_blind) - blind_reward
-        );
-    }
-    else if (timer > FRAMES(20))
-    {
-        tte_erase_rect_wrapper(BLIND_REWARD_RECT);
-        tte_erase_rect_wrapper(BLIND_REQ_TEXT_RECT);
-        obj_hide(playing_blind_token->obj);
-        affine_background_load_palette(affine_background_gfxPal);
-        state_info[game_state].substate = BLIND_PANEL_EXIT;
-        timer = TM_ZERO;
-    }
-}
-
-static void game_round_end_panel_exit()
-{
-    // TODO: make heads or tails of what's going on here and replace
-    // magic numbers.
-    if (timer < 8)
-    {
-        main_bg_se_copy_rect_1_tile_vert(TOP_LEFT_PANEL_ANIM_RECT, SCREEN_UP);
-
-        if (timer == 1) // Copied from shop. Feels slightly too niche of a function for me
-                        // personally to make one.
-        {
-            reset_top_left_panel_bottom_row();
-        }
-        else if (timer == 2)
-        {
-            int y = 5;
-            memset16(&se_mem[MAIN_BG_SBB][32 * (y - 1)], 0x0001, 1);
-            memset16(&se_mem[MAIN_BG_SBB][1 + 32 * (y - 1)], 0x0002, 7);
-            memset16(&se_mem[MAIN_BG_SBB][8 + 32 * (y - 1)], 0x0401, 1);
-        }
-    }
-    else if (timer > FRAMES(20))
-    {
-        memset16(&pal_bg_mem[REWARD_PANEL_BORDER_PID], 0x1483, 1);
-        state_info[game_state].substate = DISPLAY_REWARDS;
-        timer = TM_ZERO;
-    }
-}
-
-static inline void game_round_end_print_separator_ellipsis(void)
-{
-    int x =
-        (ROUND_END_REWARDS_ELLIPSIS_POS.x + timer - TM_REWARDS_ELLIPSIS_PRINT_START) * TILE_SIZE;
-    int y = (ROUND_END_REWARDS_ELLIPSIS_POS.y) * TILE_SIZE;
-
-    tte_printf("#{P:%d,%d; cx:0x%X000}.", x, y, TTE_WHITE_PB);
-}
-
-// TODO: Allow for more generic rewards and consolidate with game_round_end_print_interest_reward()
-static inline void game_round_end_print_hand_reward(int hand_y_offset)
-{
-    int hand_y = ROUND_END_REWARDS_ELLIPSIS_POS.y + hand_y_offset;
-    if (timer == TM_DISPLAY_REWARDS_CONT_WAIT)
-    {
-        game_round_end_extend_black_panel_down(hand_y);
-
-        tte_printf(
-            "#{P:%d,%d; cx:0x%X000}%d #{cx:0x%X000}Hands",
-            ROUND_END_REWARD_TEXT_X,
-            hand_y * TILE_SIZE,
-            TTE_BLUE_PB,
-            hand_reward,
-            TTE_WHITE_PB
-        );
-    }
-    // Increment the hand reward text until the hand reward variable is depleted
-    else if (timer > TM_HAND_REWARD_INCR_WAIT && timer % FRAMES(TM_REWARD_INCREMENT_INTERVAL) == 0)
-    {
-        hand_reward--;
-        tte_printf(
-            "#{P:%d, %d; cx:0x%X000}$%d",
-            ROUND_END_REWARD_AMOUNT_X,
-            hand_y * TILE_SIZE,
-            TTE_YELLOW_PB,
-            hands - hand_reward
-        );
-        if (hand_reward == 0)
-        {
-            interest_start_time = timer + TM_REWARD_DISPLAY_INTERVAL;
-        }
-    }
-}
-
-static inline void game_round_end_print_interest_reward(int interest_y_offset)
-{
-    int interest_y = ROUND_END_REWARDS_ELLIPSIS_POS.y + interest_y_offset;
-
-    if (timer == interest_start_time)
-    {
-        game_round_end_extend_black_panel_down(interest_y);
-
-        tte_printf(
-            "#{P:%d,%d; cx:0x%X000}%d #{cx:0x%X000}Interest",
-            ROUND_END_REWARD_TEXT_X,
-            interest_y * TILE_SIZE,
-            TTE_YELLOW_PB,
-            interest_reward,
-            TTE_WHITE_PB
-        );
-    }
-    // Increment the interest reward text until the interest reward variable is depleted
-    else if (timer > interest_start_time + TM_REWARD_DISPLAY_INTERVAL &&
-             timer % FRAMES(TM_REWARD_INCREMENT_INTERVAL) == 0)
-    {
-        interest_to_count--;
-        tte_printf(
-            "#{P:%d, %d; cx:0x%X000}$%d",
-            ROUND_END_REWARD_AMOUNT_X,
-            interest_y * TILE_SIZE,
-            TTE_YELLOW_PB,
-            interest_reward - interest_to_count
-        );
-    }
-}
-
-static void game_round_end_display_rewards()
-{
-    int hand_y_offset = 0;
-    int interest_y_offset = 0;
-
-    if (hands > 0)
-    {
-        hand_y_offset = 1;
-    }
-    else
-    {
-        interest_start_time = TM_DISPLAY_REWARDS_CONT_WAIT;
-    }
-
-    if (interest_reward > 0)
-    {
-        interest_y_offset = hand_y_offset + 1;
-    }
-
-    // Once all rewards are accounted for go to the next state
-    if (hand_reward <= 0 && interest_to_count <= 0)
-    {
-        timer = TM_ZERO;
-        state_info[game_state].substate = DISPLAY_CASHOUT;
-    }
-    else if (timer == TM_START_ROUND_END_REWARDS_ANIM)
-    {
-        game_round_end_extend_black_panel_down(ROUND_END_REWARDS_ELLIPSIS_POS.y);
-    }
-    else if (timer < TM_REWARDS_ELLIPSIS_PRINT_END)
-    {
-        game_round_end_print_separator_ellipsis();
-    }
-    else if (timer >= TM_DISPLAY_REWARDS_CONT_WAIT && hand_reward > 0)
-    {
-        game_round_end_print_hand_reward(hand_y_offset);
-    }
-    else if (interest_start_time != UNDEFINED && timer >= interest_start_time &&
-             interest_to_count > 0)
-    {
-        game_round_end_print_interest_reward(interest_y_offset);
-    }
-}
-
-static inline void game_round_end_cashout(void)
-{
-    // Reward the player
-    money += hands + blind_get_reward(current_blind) + calculate_interest_reward();
-    display_money();
-
-    hands = max_hands;          // Reset the hands to the maximum
-    discards = max_discards;    // Reset the discards to the maximum
-    display_hands(hands);       // Set the hands display
-    display_discards(discards); // Set the discards display
-
-    score = 0;
-    display_score(score); // Set the score display
-}
-
-static void game_round_end_display_cashout()
-{
-    if (timer == FRAMES(40))
-    {
-        // Put the "cash out" button onto the round end panel
-        main_bg_se_copy_expand_3x3_rect(CASHOUT_DEST_RECT, CASHOUT_SRC_3X3_RECT_POS);
-
-        int cashout_amount = hands + blind_get_reward(current_blind) + calculate_interest_reward();
-
-        bool omit_space = cashout_amount >= 10;
-        tte_printf(
-            "#{P:%d, %d; cx:0x%X000}Cash Out:%s$%d",
-            CASHOUT_TEXT_RECT.left,
-            CASHOUT_TEXT_RECT.top,
-            TTE_WHITE_PB,
-            omit_space ? "" : " ",
-            cashout_amount
-        );
-    }
-
-    // Wait until the player presses A to cash out
-    else if (timer > FRAMES(40) && key_hit(SELECT_CARD))
-    {
-        game_round_end_cashout();
-
-        state_info[game_state].substate = DISMISS_ROUND_END_PANEL; // Go to the next state
-        timer = TM_ZERO;
-
-        obj_hide(round_end_blind_token->obj);          // Hide the blind token object
-        tte_erase_rect_wrapper(BLIND_TOKEN_TEXT_RECT); // Erase the blind token text
-    }
-}
-
-static void game_round_end_dismiss_round_end_panel()
-{
-    Rect round_end_down = ROUND_END_MENU_RECT;
-    round_end_down.top--;
-    main_bg_se_copy_rect_1_tile_vert(round_end_down, SCREEN_DOWN);
-
-    if (timer >= TM_DISMISS_ROUND_END_TM)
-    {
-        timer = TM_ZERO;
-        state_info[game_state].substate = ROUND_END_EXIT;
-    }
 }
 
 static Rect get_text_rect_under_sprite_object(SpriteObject* sprite_object)
